refactor: significant changes to wrapper base - delegate requests in wrapper base to non-term requests only - requests using term now delegated to different struct - use builder for requests to allow for overriding requests
