+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod raw_types;
-pub mod search;
 pub mod types;
 pub mod util;
 pub mod wrapper;
